pmm, some cleanup and fixes

+++ sdk/includes/rtl/mem.h
@@ -5,7 +5,7 @@
 
 
 void* rtl_copy_mem(const void* src, void* dest, usize len);
-void* rtl_set_mem(void* dest, usize value, usize len);
+void* rtl_set_mem(void* dest, u8 value, usize len);
 void rtl_zero_mem(void* dest, usize len);
 isize rtl_compare_mem(const void* first, const void* second, usize len);
 u32 rtl_bswap32(u32 b);
