+++ sdk/rtl/src/rtlmem.c
@@ -11,11 +11,11 @@ void* rtl_copy_mem(const void* src, void* dest, usize len){
     return destInt;
 }
 
-void* rtl_set_mem(void* dest, usize value, usize len){
-    i8* destInt = dest;
+void* rtl_set_mem(void* dest, u8 value, usize len){
+    u8* destInt = dest;
 
     for(usize i = 0; i < len; i++){
-        destInt[i] = (uint8_t)value;
+        destInt[i] = value;
     }
     return destInt;
 }
@@ -33,7 +33,7 @@ isize rtl_compare_mem(const void* first, const void* second, usize len){
         }
     }
     return 0;
-}   
+}
 
 
 u32 rtl_bswap32(u32 b){
@@ -44,7 +44,7 @@ u32 rtl_bswap32(u32 b){
 }
 
 u64 rtl_bswap64(u64 b){
-    return ((0xFF00000000000000UL & b) >> 56) | 
+    return ((0xFF00000000000000UL & b) >> 56) |
         ((0x00FF000000000000UL & b) >> 40) |
         ((0x0000FF0000000000UL & b) >> 24) |
         ((0x000000FF00000000UL & b) >> 8) |
@@ -67,10 +67,9 @@ void* memcpy(void* dest, const void* src, usize num){
 }
 
 void* memset(void* dest, int value, usize len){
-    return rtl_set_mem(dest, value, len);
+    return rtl_set_mem(dest, (u8)value, len);
 }
 
 int memcmp(const void* ptr1, const void* ptr2, usize len){
     return rtl_compare_mem(ptr1, ptr2, len);
 }
-
